backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
cryptography==41.0.3
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0

